ClientSpec split into multiple modules
